Refactor API routers into feature modules
