+++ root/app/api/misc.py
@@ -1,0 +1,12 @@
+from fastapi import APIRouter
+
+
+router = APIRouter(tags=["misc"])
+
+
+@router.get("/activity/{id}")
+async def get_activity(id: int):
+    return {"id": id, "activity": "Demo activity"}
+
+
+__all__ = ["router"]
