+++ root/app/api/users.py
@@ -1,0 +1,339 @@
+import hashlib
+import logging
+import secrets
+import smtplib
+import ssl
+from datetime import datetime, timedelta, timezone
+from email.message import EmailMessage
+from pathlib import Path
+from typing import List, Optional
+
+from fastapi import (
+    APIRouter,
+    BackgroundTasks,
+    Depends,
+    File,
+    HTTPException,
+    Query,
+    UploadFile,
+    status,
+)
+from sqlalchemy import select, update
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from app import crud
+from app.api.deps import get_current_user
+from app.api.utils import save_upload
+from app.auth.security import get_password_hash
+from app.core.config import settings
+from app.core.database import get_db
+from app.models import models
+from app.schemas import schemas
+
+logger = logging.getLogger(__name__)
+
+
+router = APIRouter()
+password_router = APIRouter(prefix="/password", tags=["password"])
+users_router = APIRouter(prefix="/users", tags=["users"])
+groups_router = APIRouter(prefix="/groups", tags=["groups"])
+
+
+def _hash_token(token: str) -> str:
+    return hashlib.sha256(token.encode()).hexdigest()
+
+
+def _redact_sensitive_query(url: str) -> str:
+    from urllib.parse import parse_qsl, urlencode, urlsplit, urlunsplit
+
+    try:
+        parts = urlsplit(url)
+    except ValueError:
+        return "[redacted]"
+    redacted_items = []
+    for key, value in parse_qsl(parts.query, keep_blank_values=True):
+        if key.lower() in {"token", "code"}:
+            redacted_items.append((key, "***redacted***"))
+        else:
+            redacted_items.append((key, value))
+    sanitized_query = urlencode(redacted_items, doseq=True)
+    sanitized = parts._replace(query=sanitized_query)
+    result = urlunsplit(sanitized)
+    return result or "[redacted]"
+
+
+def _send_reset_email(to_email: str, link: str, full_name: str = "") -> None:
+    host = settings.smtp_host or ""
+    port = int(settings.smtp_port or 0)
+    user = settings.smtp_user or ""
+    password = settings.smtp_password or ""
+    mail_from = settings.mail_from or "no-reply@example.com"
+    security = (
+        settings.smtp_security or ("starttls" if settings.smtp_starttls else "none")
+    ).lower()
+    name = f", {full_name}" if full_name else ""
+    html = f"""
+    <div style="font-family:Inter,Arial,sans-serif">
+      <h2>Сброс пароля</h2>
+      <p>Здравствуйте{name}!</p>
+      <p>Вы запросили сброс пароля в Экосистеме ГУУ. Ссылка действует 45 минут.</p>
+      <p><a href="{link}" style="display:inline-block;padding:10px 16px;background:#1d5fff;color:#fff;border-radius:8px;text-decoration:none">Сбросить пароль</a></p>
+      <p>Если вы не запрашивали сброс, проигнорируйте это письмо.</p>
+    </div>
+    """
+    msg = EmailMessage()
+    msg["Subject"] = "Сброс пароля — Экосистема ГУУ"
+    msg["From"] = mail_from
+    msg["To"] = to_email
+    msg.set_content(f"Ссылка для сброса пароля: {link}\nОна действует 45 минут.")
+    msg.add_alternative(html, subtype="html")
+    try:
+        if not host or not port:
+            safe_link = _redact_sensitive_query(link)
+            logger.warning(
+                "password.reset_email.fallback",
+                extra={"email": to_email, "link": safe_link},
+            )
+            return
+        ctx = ssl.create_default_context()
+        if security == "ssl":
+            with smtplib.SMTP_SSL(host, port, context=ctx, timeout=10) as s:
+                if user:
+                    s.login(user, password)
+                s.send_message(msg)
+        elif security == "starttls":
+            with smtplib.SMTP(host, port, timeout=10) as s:
+                s.ehlo()
+                s.starttls(context=ctx)
+                s.ehlo()
+                if user:
+                    s.login(user, password)
+                s.send_message(msg)
+        else:
+            with smtplib.SMTP(host, port, timeout=10) as s:
+                if user:
+                    s.login(user, password)
+                s.send_message(msg)
+    except Exception:
+        safe_link = _redact_sensitive_query(link)
+        logger.error(
+            "password.reset_email.error",
+            extra={"email": to_email, "link": safe_link},
+            exc_info=True,
+        )
+
+
+@password_router.post("/forgot")
+async def forgot_password(
+    payload: schemas.ForgotPasswordIn,
+    bg: BackgroundTasks,
+    db: AsyncSession = Depends(get_db),
+):
+    result = await db.execute(
+        select(models.User).where(models.User.email == payload.email)
+    )
+    user = result.scalar_one_or_none()
+    if user:
+        token = secrets.token_urlsafe(32)
+        token_hash = _hash_token(token)
+        expires = datetime.now(timezone.utc) + timedelta(minutes=45)
+        db.add(
+            models.PasswordResetToken(
+                user_id=user.id, token_hash=token_hash, expires_at=expires, used=False
+            )
+        )
+        await db.commit()
+        base = settings.app_base_url_clean
+        reset_link = f"{base}/reset-password?token={token}"
+        bg.add_task(_send_reset_email, user.email, reset_link, user.full_name or "")
+    return {"ok": True}
+
+
+@password_router.post("/reset")
+async def reset_password(
+    payload: schemas.ResetPasswordIn, db: AsyncSession = Depends(get_db)
+):
+    token_hash = _hash_token(payload.token)
+    result = await db.execute(
+        select(models.PasswordResetToken).where(
+            models.PasswordResetToken.token_hash == token_hash,
+            models.PasswordResetToken.used.is_(False),  # E712 -> .is_(False)
+        )
+    )
+    rec = result.scalar_one_or_none()
+    if not rec or rec.expires_at < datetime.now(timezone.utc):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Недействительная или просроченная ссылка",
+        )
+    user = await db.get(models.User, rec.user_id)
+    if not user or not getattr(user, "is_active", True):
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Недействительная ссылка"
+        )
+    try:
+        user.hashed_password = get_password_hash(payload.password)
+    except ValueError as exc:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc))
+    rec.used = True
+    await db.execute(
+        update(models.PasswordResetToken)
+        .where(
+            models.PasswordResetToken.user_id == rec.user_id,
+            models.PasswordResetToken.used.is_(False),  # E712 -> .is_(False)
+        )
+        .values(used=True)
+    )
+    await db.commit()
+    return {"ok": True}
+
+
+@users_router.get("/me", response_model=schemas.UserOut)
+async def me(user: models.User = Depends(get_current_user)):
+    return user
+
+
+@users_router.put("/me", response_model=schemas.UserOut)
+async def update_me(
+    data: schemas.UserProfileUpdate,
+    db: AsyncSession = Depends(get_db),
+    user: models.User = Depends(get_current_user),
+):
+    db_user = await db.get(models.User, user.id)
+    for field, value in data.model_dump(exclude_unset=True).items():
+        setattr(db_user, field, value)
+    await db.commit()
+    await db.refresh(db_user)
+    return db_user
+
+
+@users_router.post("/me/avatar", response_model=schemas.UserOut)
+async def upload_avatar(
+    file: UploadFile = File(...),
+    db: AsyncSession = Depends(get_db),
+    user: models.User = Depends(get_current_user),
+):
+    url = await save_upload(file, "avatars", f"user_{user.id}_avatar")
+    db_user = await db.get(models.User, user.id)
+    db_user.avatar_url = url
+    await db.commit()
+    await db.refresh(db_user)
+    return db_user
+
+
+@users_router.post("/me/cover", response_model=schemas.UserOut)
+async def upload_cover(
+    file: UploadFile = File(...),
+    db: AsyncSession = Depends(get_db),
+    user: models.User = Depends(get_current_user),
+):
+    url = await save_upload(file, "covers", f"user_{user.id}_cover")
+    db_user = await db.get(models.User, user.id)
+    db_user.cover_url = url
+    await db.commit()
+    await db.refresh(db_user)
+    return db_user
+
+
+@users_router.post("", response_model=schemas.UserOut)
+async def create_user(data: schemas.UserCreate, db: AsyncSession = Depends(get_db)):
+    if data.role in ["teacher", "admin"]:
+        if not data.invite_code:
+            raise HTTPException(
+                status_code=400,
+                detail="Необходим уникальный код для регистрации преподавателя/админа",
+            )
+        q = select(models.InviteCode).where(
+            models.InviteCode.code == data.invite_code,
+            models.InviteCode.role == data.role,
+            models.InviteCode.is_active.is_(True),  # E712 -> .is_(True)
+        )
+        code_obj = (await db.execute(q)).scalar_one_or_none()
+        if not code_obj:
+            raise HTTPException(status_code=400, detail="Неверный или неактивный код")
+    user = await crud.create_user(db, data)
+    return user
+
+
+@users_router.get("", response_model=List[schemas.UserOut])
+async def get_users(
+    db: AsyncSession = Depends(get_db),
+    full_name: Optional[str] = Query(None),
+    group_id: Optional[int] = Query(None),
+    role: Optional[str] = Query(None),
+    user: models.User = Depends(get_current_user),
+):
+    if user.role != "admin":
+        raise HTTPException(status_code=403, detail="forbidden")
+    return await crud.get_users(db, full_name=full_name, group_id=group_id, role=role)
+
+
+@users_router.patch("/{user_id}", response_model=schemas.UserOut)
+async def update_user_admin(
+    user_id: int,
+    data: schemas.UserAdminUpdate,
+    db: AsyncSession = Depends(get_db),
+    user: models.User = Depends(get_current_user),
+):
+    if user.role != "admin":
+        raise HTTPException(status_code=403, detail="forbidden")
+    return await crud.admin_update_user(db, user_id, data)
+
+
+@users_router.delete("/me/avatar", response_model=schemas.UserOut)
+async def delete_avatar(
+    db: AsyncSession = Depends(get_db), user: models.User = Depends(get_current_user)
+):
+    db_user = await db.get(models.User, user.id)
+    if db_user.avatar_url:
+        base_dir = settings.static_dir_path
+        rel_path = db_user.avatar_url.replace("/static/", "", 1).lstrip("/")
+        avatar_path = base_dir / Path(rel_path)
+        if avatar_path.exists():
+            try:
+                avatar_path.unlink()
+            except Exception:
+                pass
+    db_user.avatar_url = None
+    await db.commit()
+    await db.refresh(db_user)
+    return db_user
+
+
+@users_router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_user(
+    user_id: int,
+    db: AsyncSession = Depends(get_db),
+    user: models.User = Depends(get_current_user),
+):
+    if user.role != "admin":
+        raise HTTPException(status_code=403, detail="forbidden")
+    if user.id == user_id:
+        raise HTTPException(status_code=400, detail="Нельзя удалить самого себя")
+    await crud.delete_user(db, user_id)
+    return None
+
+
+@groups_router.post("", response_model=schemas.GroupOut)
+async def create_group(
+    data: schemas.GroupCreate,
+    db: AsyncSession = Depends(get_db),
+    user: models.User = Depends(get_current_user),
+):
+    if user.role != "admin":
+        raise HTTPException(status_code=403, detail="forbidden")
+    return await crud.create_group(db, data)
+
+
+@groups_router.get("", response_model=List[schemas.GroupOut])
+async def get_groups(db: AsyncSession = Depends(get_db)):
+    result = await db.execute(select(models.Group))
+    return result.scalars().all()
+
+
+router.include_router(password_router)
+router.include_router(users_router)
+router.include_router(groups_router)
+
+
+__all__ = ["router"]
