+++ root/app/api/utils.py
@@ -1,0 +1,9 @@
+from fastapi import UploadFile
+
+from app.utils.files import save_image
+
+
+async def save_upload(file: UploadFile, subdir: str, prefix: str) -> str:
+    """Persist an uploaded file using the shared image helper."""
+
+    return await save_image(file, subdir, prefix)
