+++ root/app/main.py
@@ -7,8 +7,12 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 from sqlalchemy import text
 
+from app.api.events import router as events_router
+from app.api.misc import router as misc_router
+from app.api.news import router as news_router
 from app.api.notifications import router as notifications_router
-from app.api.routes import router as main_router
+from app.api.schedule import router as schedule_api_router
+from app.api.users import router as users_router
 from app.api.spotify import router as spotify_router
 from app.auth.auth import router as auth_router
 from app.core.config import settings
@@ -133,4 +137,8 @@ app.include_router(notifications_router)
 app.include_router(push_router)
 app.include_router(legacy_push_router)
 app.include_router(schedule_router)
-app.include_router(main_router)
+app.include_router(users_router)
+app.include_router(events_router)
+app.include_router(news_router)
+app.include_router(schedule_api_router)
+app.include_router(misc_router)
